+++ script/helper/copy-repo-files.ts
@@ -100,8 +100,12 @@ for (const config of options.config) {
                         console.log(`copy file from: ${src}`);
                         console.log(`to: ${dest}`);
                     } else {
-                        fs.unlinkSync(dest);
-                        console.log(`file removed: ${dest}`);
+                        try {
+                            fs.unlinkSync(dest);
+                            console.log(`file removed: ${dest}`);
+                        } catch {
+                            console.error(`error removing file (ignored): ${dest}`);
+                        }
                     }
                 }
             });